pktgen: split software / hardware header files
