+++ ikernels/hls/pktgen.hpp
@@ -36,74 +36,4 @@
 
 #define PKTGEN_SCHEDULER 0x20
 
-#include <ikernel.hpp>
-#include <gateway.hpp>
-#include <ntl/context_manager.hpp>
-#include <ntl/scheduler.hpp>
-
-DECLARE_TOP_FUNCTION(pktgen_top);
-
-struct pktgen_context {
-    pktgen_context() : cur_packet(0), burst_size(0), metadata()
-    {}
-
-    /** Number of the current packet in its burst */
-    uint32_t cur_packet;
-    /** Size of each burst in packets */
-    uint32_t burst_size;
-    /** Length of data in the data array (in elements) */
-    uint32_t data_length;
-    /** Packet metadata */
-    hls_ik::metadata metadata;
-};
-
-class pktgen_contexts : public ntl::context_manager<pktgen_context, LOG_NUM_IKERNELS>
-{
-public:
-    int rpc(int address, int *value, hls_ik::ikernel_id_t ikernel_id, bool read);
-};
-
-class pktgen : public hls_ik::ikernel {
-public:
-    pktgen();
-
-    void step(hls_ik::ports& ports, hls_ik::tc_ikernel_data_counts& tc);
-
-    int reg_write(int address, int value, hls_ik::ikernel_id_t ikernel_id);
-    int reg_read(int address, int* value, hls_ik::ikernel_id_t ikernel_id);
-
-private:
-    ntl::scheduler<LOG_NUM_IKERNELS> sched;
-
-    void data_plane(hls_ik::pipeline_ports& p, hls_ik::tc_pipeline_data_counts& tc);
-    void sched_wrapper();
-    void check_quantum_complete(hls_ik::pipeline_ports& p, hls_ik::tc_pipeline_data_counts& tc);
-
-    enum { IDLE, INPUT_PACKET, DUPLICATE } state;
-    /** The metadata for newly incoming packet */
-    hls_ik::metadata metadata;
-    /** The context of the current transmitting ikernel */
-    pktgen_context context;
-    /** Context updates from data plane */
-    hls::stream<pktgen_context> context_updates;
-    /** Transmit commands from the scheduler unit */
-    hls::stream<std::tuple<pktgen_context, uint32_t> > transmit_commands;
-    /** The offset into the data array we are currently accessing */
-    size_t data_offset;
-    /** Quota for the data plane to use before context switch */
-    uint32_t quota;
-    /** Size of the data array (in elements) */
-    static const int data_size = 2048 / 32;
-    /** An array that holds the packet payload to duplicate */
-    hls_ik::axi_data data[1 << LOG_NUM_IKERNELS][data_size];
-    /** Context manager class to access per ikernel context */
-    pktgen_contexts contexts;
-
-    /* Scheduler state */
-    enum { FIND_NEXT_IKERNEL, TRANSMIT_COMMAND } scheduler_state;
-    /** The ikernel we are currently processing */
-    hls_ik::ikernel_id_t sched_ikernel_id;
-    /** Current quota from the scheduler to the data plane */
-    uint32_t quota_from_sched;
-};
 #endif
